Flag to delay the watcher for some editors
Some editors, such as Emacs, make backups while saving. This causes
files to be momentarily empty before the new contents are written. This
then results in a race between Brunch and the editor, trying to compile
the file and write the file at the same time. See also issue #971.

As a workaround, it's possible to delay the watch. This adds a new
watcher config option `awaitWriteFinish`. When true, the watcher will
wait up to 50ms for the write to stabilize.