Handle nested dependencies
This is going to allow to use v1 of some library in the app code, while
retaining the ability for some other dependency to rely on v2.

It works by generating module names like this for nested modules:
"some_dep/node_modules/conflicting_dep". In require, it tries to prepend
that prefix. If the require fails, it drops it and tries to require just
"dep".