Add support for module extensions
This is dependent on
https://github.com/brunch/commonjs-require-definition/pull/12.

This PR will allow the use of extensions in requiring module names such
as `require(‘./my-widget.html’)`. It is just the minimum change to
allow this in the app code.

With the commonjs module change, `deppack/modules.js` might be made
simpler and more robust. The `package.json` “main” definition could use
aliases. Modules can be safely referenced with or without their
extensions. But I don’t feel comfortable making the changes required in
there yet.