Rework the workers implementation
New workers are conceptually simpler — they only process the compile and
optimize jobs and don't spawn Brunch fully and process only a specific
file extension. Instead, all the compile/optimize work is parallelized.

It can be triggered with the `-J 2` flag (where instead of `2`, there
can be any number of workers), instead of hard-coding that into the
config file. Config file is not the best place for this since the number
of workers is generally machine dependent, and sometimes you would want
to experiment with workers on/off or a different number of workers — and
CLI flags make it easier.