Allow non-JS compilers to optionally emit JS modules
This is useful for when you want to expose to JS some additional data
about a stylesheet, like the CSS Modules name mappings, for example.

Internally, this introduces a notion of "targets" for SourceFile, where
a single SourceFile can have up to one target of each file type (js,
style, template).