Try to recover from MODULE_NOT_FOUND in the config
If the package being required is in package.json but still can't be
loaded, that's a sign that an npm install is due, which this commit
introduces.

Note however that it does assume '.' as the root path, but given that
config's root can't be any place else, it should be ok.