Normalize paths to always use UNIX slashes internally. Ref GH-1281
Allows to get rid of platform-specific path handling in config regexps,
et cetera, and also relieves plugins from having to care about that.

Filesystem access functions work just fine with the `/`s.

Paths consistency is acheieved by normalizing the paths that come from
the watcher and by providing a local drop-in replacement for the `path`
module which always returns paths with the `/`s.