Make watcher restarts clean
This allows to cleanly restart the watcher. The previous approach with
manually trying to clean require.cache didn't work well and produced
various memory leaks.

This commit changes the brunch init process: the binary does not
directly run brunch commands, but instead, creates a child process that
processed the commands. In case the child process needs to restart, it
kills it and runs another one.

It works properly under two conditions: both global brunch binary and
the local brunch installation contain this commit. If the global brunch
is older, or if the local brunch is older, the old behavior will be
preserved.